+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 **Release Date**: 4th May 2022
 ### Features
 1. New Look and Feel for Validation Panel
-2. Add Network Load Balancer Resource to OKIT.
+2. New Resources
+    1. Network Load Balancer.
+    2. Oracle Digital Assistant (ODA Instance)
 3. New Properties layout for Object Storage Bucket implementing additional options for Auto Tiering, versioning and object events.
 ### Bug Fixes
 1. Resolve issues where match_criteria was not being tested as defined in jinja2 template for DRG Distributions